OTel.API.Context.Core.Internal

Disclaimer

In general, changes to this module will not be reflected in the library's version updates. Direct use of this module should be done with utmost care, otherwise invariants will easily be violated.

newtype Context

Constructors

Context

Fields

unContext :: Vault

emptyContext :: Context

lookupContext :: ContextKey a -> Context -> Maybe a

insertContext :: ContextKey a -> a -> Context -> Context

data ContextKey a

Constructors

ContextKey

Fields

contextKeyDebugName :: Text

contextKeyVaultKey :: Key a

contextKeyName :: ContextKey a -> Text

unsafeNewContextKey :: forall m a. MonadIO m => Text -> m (ContextKey a)

attachContextValueUsing :: forall m a b. (MonadIO m, MonadMask m) => ContextBackend a -> a -> m b -> m b

getAttachedContextValueUsing :: forall m a. (MonadIO m, MonadThrow m) => ContextBackend a -> m (Maybe a)

getAttachedContextUsing :: forall m a. (MonadIO m, MonadThrow m) => ContextBackend a -> m Context

data ContextBackend a

Constructors

ContextBackend

Fields

contextBackendStore :: Store Context

contextBackendValueKey :: ContextKey a

contextBackendRegistry :: ContextBackendRegistry

unsafeNewContextBackend :: forall m a. (MonadIO m, Typeable a) => m (ContextBackend a)

data SomeContextBackend where

Constructors

SomeContextBackend :: ContextBackend a -> SomeContextBackend

newtype ContextBackendRegistry

Constructors

ContextBackendRegistry

Fields

unContextBackendRegistry :: IORef (HashMap Unique SomeContextBackend)

emptyContextBackendRegistry :: IO ContextBackendRegistry

registerContextBackend :: Unique -> ContextBackend a -> ContextBackendRegistry -> IO ()

defaultContextBackendRegistry :: ContextBackendRegistry